main: Listen for keypresses to set the note
Spins through the notes to check if one of the note keys are being
pressed. Currently only one note can be set at a time. Due to the way
that the notes are processed, if two note keys are pressed at the same
time, then the one with a higher frequency will take precedence.

The note that is set is then picked up by the `AudioInputCallback`.