main: Fix square wave calculation
The square wave was incorrectly calcluated. If the sample is greater
than zero, then we will set the volume to the max volume. Otherwise the
value is zero.

Sine the step and the step size are the same between the sine wave and
the square wave, they were pulled out into the common scope shared
between the two. The samples happen to be the same here, but for other
waves, they will not be.